Add `useSyncExtensions` hook for syncing external state.
Keep compartmentalized extensions in sync with external state.

If the state that determines the value of a compartment necessarily lives
outside the CodeMirror EditorState (say, an app-wide theme picker), this hook
can be used to keep it in sync with the EditorState.

To avoid state tearing, this hook calls the `setEditorState` argument in the
render phase. This means that it _must_ be used in the component that owns the
EditorState. If your EditorState lives in a global state manager, you should not
use this hook.